reorganize directory structure to be domain-based
